handle -Wconversion originating in src/*.[ch]
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/matrix@4814 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/Matrix/src/expm.c
@@ -68,8 +68,8 @@ SEXP geMatrix_expm(SEXP x)
     }
 
     /* Pade' approximation. Powers v^8, v^7, ..., v^1 */
-    memset(npp, 0, sizeof(double) * nsqr);
-    memset(dpp, 0, sizeof(double) * nsqr);
+    memset(npp, 0, sizeof(double) * (size_t) nsqr);
+    memset(dpp, 0, sizeof(double) * (size_t) nsqr);
     m1_j = -1;
     for (j = 7; j >=0; j--) {
 	double mult = padec[j];
